Renamed file_ostream.hpp to file_stream.hpp
