+++ src/base/DataAllocator.cpp
@@ -13,7 +13,7 @@
 
 #include "base/DataAllocator.h"
 
-#include "io/file_ostream.hpp"
+#include "io/file_stream.hpp"
 #include "debug/macros.h"
 
 // When debugging we allocate more to
